# api-gateway/docs/swagger.yaml
basePath: /
consumes:
- application/json
definitions:
  Article:
    properties:
      content:
        type: string
        x-go-name: Content
      created_at:
        format: int32
        type: integer
        x-go-name: CreatedAt
      created_by:
        format: int32
        type: integer
        x-go-name: CreatedBy
      id:
        format: int32
        type: integer
        x-go-name: Id
      point:
        format: int32
        type: integer
        x-go-name: Point
      title:
        type: string
        x-go-name: Title
      updated_at:
        format: int32
        type: integer
        x-go-name: UpdatedAt
      view:
        format: int32
        type: integer
        x-go-name: View
    type: object
    x-go-package: github.com/maslow123/api-gateway/pkg/articles/pb
...
  GenericResponse:
    properties:
      article:
        $ref: '#/definitions/Article'
      error:
        type: string
        x-go-name: Error
      status:
        format: int64
        type: integer
        x-go-name: Status
    title: GenericResponse represents body of Register response.
    type: object
    x-go-package: github.com/maslow123/api-gateway/pkg/articles/pb
...
  SetArticlePointRequestBody:
    properties:
      article_id:
        format: int32
        type: integer
        x-go-name: ArticleID
      point:
        format: int32
        type: integer
        x-go-name: Point
    type: object
    x-go-package: github.com/maslow123/api-gateway/pkg/users/routes
...
host: http://localhost:8000
info:
  description: Documentation of BuddyAPI.
  title: BuddyAPI.
  version: 1.0.0
paths:
...
  /users/set-point:
    put:
      description: For set article point based on ArticleID
      operationId: SetArticlePoint
      parameters:
      - in: body
        name: SetArticlePointPointRequest
        required: true
        schema:
          $ref: '#/definitions/SetArticlePointRequestBody'
      responses:
        "200":
          $ref: '#/responses/SetArticlePoint'
      tags:
      - Users
produces:
- application/json
responses:
...
  SetArticlePoint:
    description: Response for set article point
    schema:
      $ref: '#/definitions/GenericResponse'
...
schemes:
- http
security:
- key:
  - '[]'
securityDefinitions:
  key:
    in: header
    name: jwt-token
    type: apiKey
swagger: "2.0"